Search nested scopes for URL value. This allows use in Jekyll loops.
Prior to this commit, a Liquid exception like the following would be
produced:

```
Liquid Exception: undefined method `split' for nil:NilClass in [...]
```

when a template included code such as the following:

```
{% for cal in site.calendars %}
{% assign url = cal.icalendar_address %}
{% ical url: url %}{% endical %}
{% endfor %}
```

This was caused by the limitation of searching only the first scope in
the `context.scopes` array. This commit addresses the issue by iterating
over all local scopes.